Added tests for submission application.yaml validation
